Move route definition to separate class
